Collect all files from my studies in one repo
Courses are divided into their semesters, and files related to the paper
I wrote and my masters' thesis have their own directories.

Note that the commit date on this makes no sense, since this is really a
collection of several git repositories, with their own histories. Since
that history was not useful to keep around, and having a single repo
makes more sense for this stuff in the future, they now live together in
this one.